# docs/swagger.yaml
basePath: /api/v1
definitions:
  Response:
    properties:
      code:
        type: integer
      data: {}
      message: {}
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  model.CreateFile:
    properties:
      cover:
        type: string
      kind:
        type: string
      name:
        type: string
      number:
        type: integer
    type: object
...
  model.DriftingPicture:
    properties:
      cover:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      kind:
        type: integer
      name:
        type: string
      number:
        type: integer
      ownerID:
        type: integer
      theme:
        type: string
      updatedAt:
        type: string
      writer_number:
        type: integer
    type: object
  model.GetFileId:
    properties:
      id:
        type: integer
    type: object
...
  model.Invite:
    properties:
      file_id:
        type: integer
      file_kind:
        type: string
      friend_id:
        type: integer
      host_id:
        type: integer
    required:
    - file_id
    - file_kind
    type: object
  model.JoinedDrifting:
    properties:
      DriftingNovelID:
        type: integer
      driftingDrawingID:
        type: integer
      driftingNoteID:
        type: integer
      driftingPictureID:
        type: integer
      studentID:
        type: integer
    type: object
...
host: 116.204.121.9:61583
info:
  contact:
    email: SJMbaiyang@163.com
    name: KitZhangYs
  description: 漂流API
  termsOfService: http://swagger.io/terms/
  title: Drifting API
  version: "1.0"
paths:
...
  /api/v1/drifting_picture/accept:
    post:
      consumes:
      - application/json
      description: 接受好友创作邀请，注：该接口仅负责删除对应邀请记录，后续操作需调用参与创作接口
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 要通过的邀请
        in: body
        name: TheInvite
        required: true
        schema:
          $ref: '#/definitions/model.Invite'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"Success"}'
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: '{"message":"Failure"}'
          schema:
            $ref: '#/definitions/Response'
      summary: 接受创作邀请
      tags:
      - driftingpicture
  /api/v1/drifting_picture/create:
    get:
      consumes:
      - application/json
      description: 获取对应用户创建的漂流照片
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"获取成功"}'
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: '{"message":"Failure"}'
          schema:
            $ref: '#/definitions/Response'
      summary: 获取用户漂流照片
      tags:
      - driftingpicture
    post:
      consumes:
      - application/json
      description: 创建漂流照片
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 新建漂流照片
        in: body
        name: NewPicture
        required: true
        schema:
          $ref: '#/definitions/model.CreateFile'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"创建成功"}'
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: '{"message":"Failure"}'
          schema:
            $ref: '#/definitions/Response'
      summary: 创建漂流照片
      tags:
      - driftingpicture
  /api/v1/drifting_picture/delete:
    delete:
      consumes:
      - application/json
      description: 删除指定漂流相片
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 要删除的漂流相片
        in: body
        name: ThePicture
        required: true
        schema:
          $ref: '#/definitions/model.DriftingPicture'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"删除成功"}'
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: '{"message":"删除失败，您有可能不是该文件的主人，或者该文件不存在"}'
          schema:
            $ref: '#/definitions/Response'
      summary: 删除漂流相片
      tags:
      - driftingpicture
  /api/v1/drifting_picture/detail:
    post:
      consumes:
      - application/json
      description: 获取漂流本内容，需在json中提供漂流照片的ID
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 获取的ID
        in: body
        name: FDriftingNote
        required: true
        schema:
          $ref: '#/definitions/model.GetFileId'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"获取成功"}'
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: '{"message":"获取失败"}'
          schema:
            $ref: '#/definitions/Response'
      summary: 获取漂流照片内容
      tags:
      - driftingpicture
  /api/v1/drifting_picture/draw:
    post:
      consumes:
      - application/json
      description: 创作漂流照片
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 内容
        in: formData
        name: picture
        required: true
        type: file
      - description: id
        in: formData
        name: file_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"创建成功"}'
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: '{"message":"创建失败"}'
          schema:
            $ref: '#/definitions/Response'
      summary: 创作漂流照片
      tags:
      - driftingpicture
  /api/v1/drifting_picture/invite:
    get:
      consumes:
      - application/json
      description: 获取该用户的邀请信息
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"获取成功"}'
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: '{"message":"获取信息失败"}'
          schema:
            $ref: '#/definitions/Response'
      summary: 获取邀请信息
      tags:
      - driftingpicture
    post:
      consumes:
      - application/json
      description: 邀请好友创作，需在json中提供好友学号，漂流本ID，及文件类型(漂流本需注明是DriftingNote)
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 新建邀请
        in: body
        name: NewInvite
        required: true
        schema:
          $ref: '#/definitions/model.Invite'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"邀请成功"}'
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: '{"message":"邀请失败，你可能已邀请过该好友"}'
          schema:
            $ref: '#/definitions/Response'
      summary: 邀请好友进行创作
      tags:
      - driftingpicture
  /api/v1/drifting_picture/join:
    get:
      consumes:
      - application/json
      description: 获取用户参加得漂流照片信息
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"获取成功"}'
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: '{"message":"Failure"}'
          schema:
            $ref: '#/definitions/Response'
      summary: 获取用户参加的漂流照片信息
      tags:
      - driftingpicture
    post:
      consumes:
      - application/json
      description: 参加漂流照片创作(仅参加)
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 要参加的漂流照片
        in: body
        name: Joining
        required: true
        schema:
          $ref: '#/definitions/model.JoinedDrifting'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"Success"}'
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: '{"message":"Failure"}'
          schema:
            $ref: '#/definitions/Response'
      summary: 参加漂流照片创作(仅参加)
      tags:
      - driftingpicture
  /api/v1/drifting_picture/recommendation:
    get:
      consumes:
      - application/json
      description: 随机推荐一个漂流照片
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"获取成功"}'
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: '{"message":"获取失败"}'
          schema:
            $ref: '#/definitions/Response'
      summary: 随机推荐漂流照片
      tags:
      - driftingpicture
  /api/v1/drifting_picture/refuse:
    post:
      consumes:
      - application/json
      description: 拒绝创作邀请
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 拒绝邀请
        in: body
        name: TheInvite
        required: true
        schema:
          $ref: '#/definitions/model.Invite'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"拒绝成功"}'
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: '{"message":"拒绝失败"}'
          schema:
            $ref: '#/definitions/Response'
      summary: 拒绝邀请
      tags:
      - driftingpicture
...
swagger: "2.0"